push logic complete

+++ YuePaoBa/Class/Account/Model/YPBRegisterModel.m
@@ -54,8 +54,8 @@
         }
     }
     
-    NSString *heightArea = [NSString stringWithFormat:@"%ld~%ld", user.targetHeight.min, user.targetHeight.max];
-    NSString *ageArea = [NSString stringWithFormat:@"%ld~%ld", user.targetAge.min, user.targetAge.max];
+    NSString *heightArea = [NSString stringWithFormat:@"%ld~%ld", (long)user.targetHeight.min, (long)user.targetHeight.max];
+    NSString *ageArea = [NSString stringWithFormat:@"%ld~%ld", (long)user.targetAge.min, (long)user.targetAge.max];
     return @{@"sex":user.sex,
              @"nickName":user.nickName,
              @"uuid":[YPBUtil activationId],
